+++ INTEGRA-backend/routes/api/Marketing/lead.php
@@ -3,11 +3,12 @@
 use App\Http\Controllers\MarketingControllers\LeadController;
 use Illuminate\Support\Facades\Route;
 
+
 Route::prefix('marketing')->group(function(){
-    Route::get('/leads', [LeadController::class, 'index']);
-    Route::get('/lead/show/{id}', [LeadController::class, 'show']);
-    Route::post('/lead/store', [LeadController::class, 'store']);
-    Route::delete('/lead/destroy/{id}', [LeadController::class, 'destroy']);
-    Route::put('/lead/update/{id}', [LeadController::class, 'update']);
+    Route::get('/leads', [LeadController::class, 'index'])->middleware('permission:index lead');
+    Route::get('/lead/show/{id}', [LeadController::class, 'show'])->middleware('permission:show lead');
+    Route::post('/lead/store', [LeadController::class, 'store'])->middleware('permission:store lead');
+    Route::delete('/lead/destroy/{id}', [LeadController::class, 'destroy'])->middleware('permission:destroy lead');
+    Route::put('/lead/update/{id}', [LeadController::class, 'update'])->middleware('permission:update lead');
 
 });
