+++ INTEGRA-backend/routes/api/Marketing/pdffile.php
@@ -3,9 +3,10 @@
 use App\Http\Controllers\MarketingControllers\PDFController;
 use Illuminate\Support\Facades\Route;
 
+
 Route::prefix('marketing')->group(function(){
-    Route::get('/pdfs', [PDFController::class, 'index']);
-    Route::get('/pdf/show/{id}', [PDFController::class, 'show']);
-    Route::post('/pdf/store/{id}', [PDFController::class, 'store']);
-    Route::delete('/pdf/destroy/{id}', [PDFController::class, 'destroy']);
+    Route::get('/pdfs', [PDFController::class, 'index'])->middleware('permission:index pdf');
+    Route::get('/pdf/show/{id}', [PDFController::class, 'show'])->middleware('permission:show pdf');
+    Route::post('/pdf/store/{id}', [PDFController::class, 'store'])->middleware('permission:store pdf');
+    Route::delete('/pdf/destroy/{id}', [PDFController::class, 'destroy'])->middleware('permission:destoy pdf');
 });
